Add proper clamping
The previous method of clamping as each sample is mixed is imperfect:
when samples are added, the result isn't always larger: this means if
two samples are added (0x6000 and 0x6000) it will be clamped down
(0x7FFF), but then if a negative sample is added (say, -0x6000), the
final value will be incorrect (0x1FFF instead of 0x6000).

Instead, we now use an intermediary 'mixer buffer' which is 32-bit
instead of 16-bit. Only once all sounds are mixed do we perform
clamping.

For now, music is disabled. I'll fix this in another commit.